refactor: migrate UpgradesPage SSG to new utils
