+++ src/pages/UpgradesPage/static.ts
@@ -1,18 +1,16 @@
 import type { InferGetStaticPropsType } from "next";
 import { serverApi } from "~/server/api/root";
-import {
-  getBundledQuestMap,
-  getBundledServantMap,
-  getBundledUpgrades
-} from "~/utils/getBundles";
+import { getQuestMap } from "~/static/data/quests";
+import { getServantMap } from "~/static/data/servants";
+import { getBundledUpgrades } from "~/static/upgrades";
 import { createUpgradeFilter, createUpgradeSorter } from "./filters";
 import { formFiltersDefault } from "./filtersReducer";
 
 export const getStaticProps = async () => {
   const [upgradesList, questMap, servantMap] = await Promise.all([
     getBundledUpgrades(),
-    getBundledQuestMap(),
-    getBundledServantMap()
+    getQuestMap(),
+    getServantMap()
   ]);
 
   const sorter = createUpgradeSorter(questMap);
